Once in a while, one of the SMALLINT columns is the maximum value, which JDBC doesn't think fits into a SHORT.
git-svn-id: https://svn.code.sf.net/p/ngsgenomelibs/code/trunk@439 2b8d95ea-eb58-0410-8a55-29fbd2c437c3

+++ src/edu/usc/epigenome/genomeLibs/MethylDb/CpgIteratorMultisample.java
@@ -156,13 +156,13 @@ public class CpgIteratorMultisample implements Iterator<Cpg[]> {
 			{
 				Cpg cpg = new Cpg(curRS.getInt("cpg" + i + ".chromPos"),
 						curRS.getString("cpg" + i + ".strand").equals("-"),
-						curRS.getShort("cpg" + i + ".totalReads"),
-						curRS.getShort("cpg" + i + ".cReads"),
-						curRS.getShort("cpg" + i + ".cReadsNonconversionFilt"),
-						curRS.getShort("cpg" + i + ".tReads"),
-						curRS.getShort("cpg" + i + ".agReads"),
-						curRS.getShort("cpg" + i + ".totalReadsOpposite"),
-						curRS.getShort("cpg" + i + ".aReadsOpposite"),
+						(short)curRS.getInt("cpg" + i + ".totalReads"),
+						(short)curRS.getInt("cpg" + i + ".cReads"),
+						(short)curRS.getInt("cpg" + i + ".cReadsNonconversionFilt"),
+						(short)curRS.getInt("cpg" + i + ".tReads"),
+						(short)curRS.getInt("cpg" + i + ".agReads"),
+						(short)curRS.getInt("cpg" + i + ".totalReadsOpposite"),
+						(short)curRS.getInt("cpg" + i + ".aReadsOpposite"),
 						curRS.getInt("cpg" + i + ".cpgWeight"));
 
 				// Downsample data.
